OA-56 - parcial, modificação no tipo da resposta em IntegerAnswer e GridIntegerAnswer

+++ otus-activity/src/main/java/org/ccem/otus/model/survey/activity/filling/answer/GridIntegerAnswer.java
@@ -3,13 +3,13 @@ package org.ccem.otus.model.survey.activity.filling.answer;
 public class GridIntegerAnswer {
 	private String objectType;
 	private String customID;
-	private Integer value;
+	private Long value;
 
 	public String getCustomID() {
 		return customID;
 	}
 
-	public Integer getValue() {
+	public Long getValue() {
 		return value;
 	}
